pruebasConcepto/backend/requirements.txt
flask
kodipydent-alt
mypy